requirements.txt
SQLAlchemy==2.0.36
python-dotenv==1.0.0
bcrypt==4.1.2
PyJWT==2.8.0
click==8.1.7
rich==13.7.0
sentry-sdk==1.39.2
email-validator==2.1.0
pytest==7.4.3
pytest-cov==4.1.0